chore: Better management of the list of images in docker::images::clean

+++ roles/dotfiles/files/shell/zsh.d/docker.zsh
@@ -9,7 +9,9 @@ docker::containers::clean() {
 }
 
 docker::images::clean() {
-  docker image rm $(docker image list --filter "dangling=true" -q)
+  local -r docker_dangling_images=$(docker image list --filter "dangling=true" -q)
+
+  [[ -n "${docker_dangling_images}" ]] && docker image rm $(docker image list --filter "dangling=true" -q)
 }
 
 # autoloads
